multirom: Add FDT handling for FDT-fstab entries.
Multirom gained support for FDT-encoded fstab entries in
TODO: Link multirom commit.

In order to use the mechanism in the Multrom project,
this commit modifies the recovery part to perform the
required FDT processing when needed: It is done when the
boot images of secondary roms are injected.

The following changes were implemented:
* A flag is passed to trampoline if a secondary rom shall
  be injected. This triggers the removal of the /system
  partition entry from the FDT that otherwise would
  prevent to boot the secondary rom correctly.
* The original FDT is stored with the secondary rom such
  that it can be restored if a secondary rom shall become
  the primary rom (move or copy).
* Restore the original DTB if a secondary rom becomes the
  primary one.
* Modify the FDT if a primary rom is moved or copied to
  become a secondary rom.

Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>